requirements.txt
customtkinter
CTkMessagebox
numpy
Pillow
